move product details from 'routes' to 'directives' and start stubbing out contents including 'add to cart' button
