+++ client/shop/directives/product-detail/ProductDetail.scss
