+++ client/shop/directives/product-detail/ProductDetail.spec.js
@@ -4,8 +4,7 @@ import { SHOP_DIRECTIVES } from '../../directives';
 import { SHOP_PIPES } from '../../pipes';
 import { SHOP_COMPONENTS } from '../../routes';
 import { SHOP_PROVIDERS } from '../../providers';
-import {SHARED_IMPORTS} from '../../../shared/imports';
-import {SharedModule} from '../../../shared';
+import SHOP_IMPORTS from '../../imports';
 
 import ProductDetail from './ProductDetail';
 
@@ -15,7 +14,7 @@ describe('ProductDetail', () => {
         TestBed.configureTestingModule({
             declarations: [ ...SHOP_DIRECTIVES, ...SHOP_PIPES, ...SHOP_COMPONENTS ],
             providers: SHOP_PROVIDERS,
-            imports: [ SharedModule, ...SHARED_IMPORTS, RouterTestingModule ]
+            imports: [ ...SHOP_IMPORTS, RouterTestingModule ]
         });
     });
