+++ client/shop/directives/product-detail/index.js
@@ -1,0 +1,2 @@
+import ProductDetail from './ProductDetail';
+export default ProductDetail;
