--- client/shop/routes/product-detail/ProductDetail.html
@@ -1,4 +1,0 @@
-<h1>{{name}}</h1>
-<!--TODO: this is where child views get output, if needed
-<router-outlet></router-outlet>
--->
