--- client/shop/routes/product-detail/index.js
@@ -1,12 +1,0 @@
-import ProductDetail from './ProductDetail';
-
-/**
- * @see https://angular.io/docs/ts/latest/api/router/index/Route-interface.html
- * @type {*}
- */
-const ROUTE = {
-    path: 'shop/product/:productId',
-    component: ProductDetail
-};
-
-export default ROUTE;
